Implementacion de get desde api

+++ app/src/main/java/com/example/titancake/ui/viewmodel/CartViewModel.kt
@@ -54,7 +54,7 @@ class CartViewModel : ViewModel() {
     }
 
     // Calcula el total a pagar sumando el precio por cantidad de cada producto.
-    fun total(): Double = _carrito.value.sumOf { it.producto.precio * it.cantidad }
+    fun total(): Int = _carrito.value.sumOf { it.producto.precio * it.cantidad }
 
 }
